refactor: numbergenerator - module패키지로 이동
