+++ kotlin-baseball/src/test/kotlin/baseball/module/ComputerGeneratorTest.kt
@@ -1,9 +1,9 @@
-package baseball
+package baseball.module
 
+import baseball.main
 import camp.nextstep.edu.missionutils.test.NsTest
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
-import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
